requirement.txt
requests
beautifulsoup4
python-dotenv
tqdm
chromadb
numpy
pydantic
faiss-cpu
httpx
sentence-transformers
